Cast Streaming for desktop: Don't send audio offer/stream if no audio.
This patch adds in the necessary plumbing, when launching desktop
mirroring via the Media Router UI, to conditionally enable the audio
stream in the Mirroring Service. It also fixes a bug in the Desktop
Media Picker controller, in the case where the dialog is being auto-
bypassed: The logic in the controller was not setting the |audio_share|
property in the result (normally, the UI dialog code would have set
this).

Within the Media Router and Cast MRP, it is necessary to specify the
audio choice in the Media Router's desktop source URN since that alone
is used to turn audio capture on/off in the Mirroring Service when
launching desktop mirroring. In addition, the code has been tweaked to
clarify the distinction between an "unchosen desktop" URN versus the
specific one the user chose from the dialog.

Bug: 1110950
Change-Id: I31d437dfadc3372dc2024beba2f8591d78e91875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347374
Commit-Queue: Yuri Wiitala <miu@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801256}