Handle toolchains with/without ..\.. in variables
The packaging script previously would generate environment variables
with lots of ..\.. entries in environment variables, which is difficult
to read and increases the odds of hitting MAX_PATH limits.
crrev.com/c/2372727 changes the packaging and this change to
setup_toolchain.py adjusts the setup code to handle both the old and new
systems.

Bug: 1120785
Change-Id: I7c3449a54cf6fba9f70f52db83d940ffd4c32573
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2370604
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Scott Graham <scottmg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801277}