Only stop the moving to center with Unadjusted Movement on windows
On Windows we grab the raw mouse movements and stop actually moving the
mouse when in unadjusted movement so moving to the center does not make
sense.  However, on ChromeOS and Linux we temporarily turn off mouse
Acceleration in the settings and keep the mouse moving like normal so
The move to center calculation should not change.

Bug: 1119468
Change-Id: Ib318627340df98f55205f4489084d88c89a7fd37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2373447
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801295}