SVGImage::PopulatePaintRecordForCurrentFrameForContainer need zoom
SVGImageForContainer::PaintImageForCurrentFrame() doesn't pass on zoom,
and will thus produce a result an image that differs from how the same
image would look if it was painted using a more direct code-path - such
as Draw().
Add a |zoom| argument to
SVGImage::PopulatePaintRecordForCurrentFrameForContainer and adjust
callers to pass a relevant zoom-factor. Re-order the order of the
arguments to match other methods that accept zoom.

Bug: 1121177
Change-Id: If2e1abe8ee3f9f29e176044b1ea3261bc300a87f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2368593
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#801303}