Check SODA paths before launching service
This CL introduces a check to verify that the SODA binary and language
pack components are downloaded before launching the speech recognition
service. In the event where the files are not yet downloaded, a generic
error message will be shown to the user indicating that speech
recognition is not available.

Bug: 1121384
Change-Id: Iff628202884869a04ce54761137aca4d40497fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2373388
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Evan Liu <evliu@google.com>
Cr-Commit-Position: refs/heads/master@{#801310}