Add Read Later omnibox star entry point.
Add a new menu triggered by the StarView which gives the option to
either bookmark the tab or save it for later. If bookmark is selected
the user flow matches previous behavior. If save to read later is
selected the tab will be saved to the reading list model. Note this
change is hidden behind the read later feature flag.

Follow-up work: Update the read later icon in the menu when new icons
are available. Update the omnibox star tooltip. Update StarView/bookmark metrics and
add metrics for read later. Update menu size.

Bug: 1117023
Change-Id: I09c6cf4973338faa272bc6ec6da06e26f0f9aa8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359550
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801449}