Simplified domains: refine hover histogram
In https://chromium-review.googlesource.com/c/chromium/src/+/2363999/
I added a histogram that records the duration of each continuous hover
over the omnibox. As mpearson@ pointed out in that CL, this loses some
signal because it continues to count edits and navigations as part of
the hover (whereas what we really care about in this field trial is
the particular interaction where the user hovers and exits the omnibox
without otherwise interacting with it). This CL modifies the histogram
to emit on mouse exit or focus, whichever comes first.

Bug: 1111078
Change-Id: I49ea3cab267cc6f813a4c1dab6c7454932013e34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367791
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801512}