Do not use borders when setting padding on tab close button.
This is a start at eliminating redundant layouts and memory allocations
during layout in the tabstrip. Instead of a border, we use
kInternalPaddingKey with a pre-allocated gfx::Insets. This is the same
approach used by ToolbarButton and will also help if tabs are switched
to using a layout manager.

Ultimately, we want to eliminate both allocation of border objets as
well as all layout invalidations during layout.

Bug: 1121681
Change-Id: Icf7ac6c93eda5ccf80639ab95b2e0c705c1c5d45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2375741
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801572}