Don't send DidStartProvisionalLoad notifications for JS urls and XSLT documents
We typically expect DidStartProvisionaLoad messages to be paired with a
DidCommitLoad when the navigation is successful. For pseudo-navigations like
XSLT processing and javascript: urls, we skip the DidCommitLoad message.
If we do not also skip the DidStartProvisionalLoad message, listeners that
expect these notifications to be paired will get confused (e.g., the
inspector debugger may leave breakpoints disabled in the new Document).

Bug: 1095920
Change-Id: I5bdd18b3f4103d5d2c52fe48060c1114efb21b0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254818
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801573}