Add support for logging UMA histogram samples from trace config
Now when histogram_samples category is set only the histograms listed
in a trace config will be monitored. If there are no histograms in a
trace config, the effect of enabling the histogram_samples category will
remain the same: all the histograms will be monitored.

Monitoring all the histograms results in a significant increase in
trace size and processing time, that's why we need a way to monitor
only a subset of histograms.

Based on the events added to trace when enabling the histogram_samples
category TBMv3 UMA metrics will be implemented.

Change-Id: If2c36ba22e222c0c1efb4415aba54edcd69b2b4c
Bug: 1119834
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363747
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Reviewed-by: Deep Roy <dproy@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Commit-Queue: Julia Semavina <nuwanda@google.com>
Cr-Commit-Position: refs/heads/master@{#801583}