Fix issue when deleting from a textfield.
Fix issue when deleting from a textfield and an emoji is present, where
the last character would get deleted. Also moves card holder name down
in the list of items as requested by UX.

The main issue, which also has other unintended side effects, is that
|reconfigureCellsForItems| is called while the text is being edited.
This causes the cell to be redrawn and the cursor to change.

By doing this call in TableViewTextEditItemDelegate as opposed to
before the text changes we resolve the issue of the last character
being deleted instead of the correct one. There is a smaller issue
that remains and is not trivial to fix: after deletion the cursor
is moved to the end of the text. This also exists in other
textfields.

(cherry picked from commit f111c544b67335239d5f1b8292668cbbe87d0d24)

Bug: 1121291
Change-Id: Id035e5b3c066801a3e156360f010380104564931
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360514
Commit-Queue: Alfonso Garza <alfonsogarza@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#800385}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2376481
Reviewed-by: Siyu An <siyua@chromium.org>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/branch-heads/4240@{#62}
Cr-Branched-From: f297677702651916bbf65e59c0d4bbd4ce57d1ee-refs/heads/master@{#800218}