[MBI] Associate RenderWidgetHost with AgentSchedulingGroup.
Add a pointer to the AgentSchedulingGroupHost in RenderWidgetHost, and
route some initial calls through it instead of the RenderProcessHost.
This is needed since in the future all widget-specific communication
will be done directly at the agent scheduling group level, and not the
process level.


Bug: 1111231
Change-Id: I65bf4769bd9572046316a9f569651bcaf8e462aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358943
Commit-Queue: Tal Pressman <talp@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801665}